refactor: start with a simple solution
